FIX: file_interface - compare between signed and unsigned

+++ src/ui/file_interface/file_interface.c
@@ -70,8 +70,8 @@ u8 file_has_changed(FILE_INTERFACE* p_file) {
     if (stat(p_file->path, &file_attribute) == 0) {
         return 0;
         
-    } else if (file_attribute.st_mtime > p_file->timestamp_last_modified) {
-        return 1;        
+    } else if ((u32)file_attribute.st_mtime > (p_file->timestamp_last_modified) {
+        return 1;
     }
     
     return 0;
